pad appropriate return values (v/A/B/S) to the size of the modulus N
Tested by test_picl_vectors for getv, getA, getB, client_getS, and
server_getS.

That test also checks getu, getM, and getK (which are supposed to return
Buffers that equal the size of the hash function). However the test
vectors for those values don't start with a 0x00 byte, so the test isn't
complete. We could use additional vectors to exercise those cases.